feat: more granular details on vies faults (#6)
### Why?

For different reasons, government databases sometimes can't respond to
whether a `TaxId` is verified or unverified. The `VerificationStatus` is
set to `Unavailable` upon such responses.

However, more granular details should be provided to the client as some
faults/errors are a result of their behaviour or should impact their
behaviour. Examples would be when the client has been blocked or
rate-limited.

### What changed?

- Verification is extended with a `data` HashMap.
- Faults are mapped into data
- Successful responses are mapped into data as well (details like
address etc)